@@ -2,8 +2,8 @@ apply plugin: 'com.android.application'
 
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.1"
+    compileSdkVersion 30
+    buildToolsVersion '30.0.0'
 
     defaultConfig {
         applicationId "com.mantz_it.wearguitartuner"
@@ -21,9 +21,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':guitartunerlibrary')
-    compile 'com.google.android.support:wearable:2.7.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':guitartunerlibrary')
+    implementation 'com.google.android.support:wearable:2.7.0'
 }
 allprojects {
     repositories {
